backend/requirements.txt
Flask
pymongo
dnspython
python-dotenv
flask-cors
